Entropy failed to give results
The justifications are :
You can’t compare across different points in time. You need to
normalize and normalizing by area is not right as you can’t
differentiate between the increase due to area decrease or that due to
the increase in entropy. You need to find a way to decouple the
increase in entropy in water injections that is due to the fact that
there is spacial dispersion and due to the fact that there is an
increase in the intensity absolute value itself. So, now I will use my
center of mass, thresholding and binary masks and area under curve
ideas to come up with a proof for the increase in rok intensity in the
cells.